# v1/metrics: Prealloc maps + add benchmark