## What changed?

This PR preallocates maps where possible in the `v1/metrics` package, which we were not doing before. This provides a small, but reliable/stable boost to the performance of several operations on the hot path for serialization of metrics, mostly from reducing allocations.

## Performance Analysis / Benchmarks

I built a benchmark for the hot path that I'm the most interested in: JSON serialization when several metrics are present.
This fully exercises most of the `v1/metrics` package, and shows the full benefit of the prealloc changes.

Here's the [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) diff of 40x runs of main (as of yesterday) versus 40x runs of this PR:

```
% benchstat main.txt pr.txt
goos: linux
goarch: amd64
pkg: github.com/open-policy-agent/opa/v1/metrics
cpu: Intel(R) Core(TM) i7-8650U CPU @ 1.90GHz
                    │  main.txt   │               pr.txt               │
                    │   sec/op    │   sec/op     vs base               │
MetricsMarshaling-8   133.3µ ± 1%   124.9µ ± 1%  -6.29% (p=0.000 n=40)

                    │   main.txt   │                pr.txt                │
                    │     B/op     │     B/op      vs base                │
MetricsMarshaling-8   42.83Ki ± 0%   37.98Ki ± 0%  -11.32% (p=0.000 n=40)

                    │  main.txt  │          pr.txt           │
                    │ allocs/op  │ allocs/op   vs base       │
MetricsMarshaling-8   562.0 ± 0%   547.0 ± 0%  -2.67% (n=40)
```

- [main.txt](https://github.com/user-attachments/files/20594462/main.txt)
- [pr.txt](https://github.com/user-attachments/files/20594463/pr.txt)